src/requirements.txt
llama-cpp-python
pygobject